+++ src/resdk/resources/__init__.py
@@ -41,7 +41,7 @@ Resource classes
 .. autoclass:: resdk.resources.DescriptorSchema
    :members:
    :inherited-members:
-   
+
 .. autoclass:: resdk.resources.AnnotationValue
    :members:
    :inherited-members:
@@ -63,6 +63,10 @@ Resource classes
    :inherited-members:
 
 .. autoclass:: resdk.resources.PredictionField
+   :members:
+   :inherited-members:
+
+.. autoclass:: resdk.resources.PredictionPreset
    :members:
    :inherited-members:
 
@@ -110,7 +114,12 @@ from .data import Data
 from .descriptor import DescriptorSchema
 from .geneset import Geneset
 from .metadata import Metadata
-from .predictions import PredictionField, PredictionGroup, PredictionValue
+from .predictions import (
+    PredictionField,
+    PredictionGroup,
+    PredictionPreset,
+    PredictionValue,
+)
 from .process import Process
 from .relation import Relation
 from .sample import Sample
@@ -128,6 +137,7 @@ __all__ = (
     "Metadata",
     "PredictionField",
     "PredictionGroup",
+    "PredictionPreset",
     "PredictionValue",
     "Sample",
     "Process",