+++ src/resdk/resources/predictions.py
@@ -1,8 +1,9 @@
 """Predictions resources."""
 
+import copy
 import logging
 from enum import Enum
-from typing import TYPE_CHECKING, NamedTuple, Optional, Type, Union
+from typing import TYPE_CHECKING, Iterable, NamedTuple, Optional, Type, Union
 
 from ..utils.decorators import assert_object_exists
 from .base import BaseResource
@@ -137,6 +138,15 @@ class PredictionField(BaseResource):
         """Set prediction group."""
         self._resource_setter(payload, PredictionGroup, "_group")
 
+    def _dehydrate_resources(self, obj):
+        """Prediction fields are serialized by id only.
+
+        For other fields use default serialization.
+        """
+        if isinstance(obj, PredictionType):
+            return obj.value
+        return super()._dehydrate_resources(obj)
+
     def __repr__(self):
         """Return user friendly string representation."""
         return f"PredictionField <path: {self.group.name}.{self.name}>"
@@ -266,3 +276,106 @@ class PredictionValue(BaseResource):
             f"PredictionValue <path: {self.field.group.name}.{self.field.name}, "
             f"value: '{self.value}'>"
         )
+
+
+class PredictionFieldSet:
+    """The set of resources."""
+
+    def __init__(self, parent: BaseResource, field_name: str):
+        """Initialize the set."""
+        self._resources = set()
+        self._field_name = field_name
+        self._parent = parent
+
+    def add(self, *resources: Iterable[Union[BaseResource, int]]):
+        """Add the resources to the set."""
+        self._resources.update(
+            self._parent._get_resource(entry, PredictionField) for entry in resources
+        )
+        self._patch()
+
+    def remove(self, *resource: Iterable[BaseResource]):
+        """Remove the resources from the set."""
+        self._resources.difference_update(resource)
+        self._patch()
+
+    def set(self, resources: Iterable[Union[BaseResource, int]], patch=True):
+        """Assign the resources to the set."""
+        self._resources = set(
+            self._parent._get_resource(entry, PredictionField) for entry in resources
+        )
+        if patch:
+            self._patch()
+
+    def clear(self):
+        """Clear the set."""
+        self._resources.clear()
+        self._patch()
+
+    def __iter__(self):
+        """Iterate over the set."""
+        return iter(self._resources)
+
+    def _patch(self):
+        """Send a list of resource ids to the server."""
+        if self._parent.id is not None:
+            self._parent.api(self._parent.id).patch(
+                {self._field_name: [item.id for item in self._resources]}
+            )
+
+    def __str__(self):
+        """Return user friendly string representation."""
+        return f"ResourceSet <{self._resources}>"
+
+
+class PredictionPreset(BaseResource):
+    """Resolwe PredictionPreset resource."""
+
+    endpoint = "prediction_preset"
+
+    READ_ONLY_FIELDS = BaseResource.READ_ONLY_FIELDS
+
+    UPDATE_PROTECTED_FIELDS = BaseResource.UPDATE_PROTECTED_FIELDS + ("contributor",)
+
+    WRITABLE_FIELDS = BaseResource.WRITABLE_FIELDS + ("name", "fields")
+
+    def __init__(self, resolwe: "Resolwe", **model_data):
+        """Initialize the instance."""
+        self.logger = logging.getLogger(__name__)
+        #: prediction fields
+        self._fields = PredictionFieldSet(self, "fields")
+        super().__init__(resolwe, **model_data)
+
+    def _update_fields(self, payload):
+        """Handle fields differently."""
+        self._original_values = copy.deepcopy(payload)
+        for field_name in self._get_resource_fields():
+            if field_name == "fields":
+                self._fields.set(payload.get(field_name, []), patch=False)
+            else:
+                setattr(self, field_name, payload.get(field_name, None))
+
+    def _dehydrate_resources(self, obj):
+        """Prediction fields are serialized by id only.
+
+        For other fields use default serialization.
+        """
+        if isinstance(obj, PredictionField):
+            return obj.id
+        if isinstance(obj, PredictionFieldSet):
+            return [entry.id for entry in obj]
+        return super()._dehydrate_resources(obj)
+
+    @property
+    def fields(self):
+        """Get fields."""
+        return self._fields
+
+    @fields.setter
+    def fields(self, values: Iterable[PredictionField]):
+        """Set fields."""
+        self._fields.set(values)
+
+    def __repr__(self):
+        """Return user friendly string representation."""
+        return f"PredictionPreset <name: {self.name}>"