+++ docs/CHANGELOG.rst
@@ -17,6 +17,8 @@ Added
 - Add a resolwe method for fetching the server module versions
 - Add support for ``Python 3.13``
 - Add ``QCTables`` class to the ``tables`` module
+- Add ``cookies`` authentication method
+- Add ``PredictionPreset`` resource
 
 Changed
 -------